+++ python/alpha93/packages/commons/src/alpha93/collections/__init__.py
@@ -1,5 +1,11 @@
 from ._mapping import MappingAccessor, MutableMappingAccessor, accessor
-from ._singleton import MutableSingletonSequence, SingletonSequence
+from ._singleton import (
+    MutableSingletonSequence,
+    SingletonList,
+    SingletonSequence,
+    SingletonTuple,
+)
+from .utils import flatmap
 
 if __debug__ and __import__("typing").TYPE_CHECKING:
     from ._mapping import MappingLike, MutableMappingLike
@@ -10,6 +16,9 @@ __all__ = (
     "MutableMappingAccessor",
     "MutableMappingLike",
     "MutableSingletonSequence",
+    "SingletonList",
     "SingletonSequence",
+    "SingletonTuple",
     "accessor",
+    "flatmap",
 )