+++ python/alpha93/packages/commons/src/alpha93/collections/utils.py
@@ -1,0 +1,9 @@
+from collections.abc import Iterable
+from functools import reduce
+
+__all__ = (
+    "flatmap",
+)
+
+def flatmap[K, V](mappings: Iterable[dict[K, V]]) -> dict[K, V]:
+    return reduce(dict.__or__, mappings)